object id von int auf long geandert
git-svn-id: https://svn.r7r.at/sepm/trunk@284 240b89b8-79f3-4156-a269-5c34a98ffd57

+++ shared/src/main/java/at/fakeroot/sepm/shared/client/serialize/ClientGeoObject.java
@@ -17,14 +17,14 @@ public class ClientGeoObject extends GeoObject implements Serializable {
 	}
 
 	/**
-	 * @param id Unique ID
+	 * @param id Object ID
 	 * @param titel ObjTitel
 	 * @param xPos (Lng=Longitute)
 	 * @param yPos (Lat=Latitiute)
 	 * @param imageUrl Pin (Marker) Image
 	 * @param tags GeoObject Tags
 	 */
-	public ClientGeoObject(int id, String titel, String imageUrl, String[] tags, double xPos, double yPos){
+	public ClientGeoObject(long id, String titel, String imageUrl, String[] tags, double xPos, double yPos){
 		super(id,titel,xPos,yPos);
 		this.imageUrl=imageUrl;
 		setTags(tags);
